pyproject.toml
[tool.poetry]
name = "obs_layer_sorsimple_mbaas"
version = "0.1.0"
description = "Capa lambda escritura SOR Mbaas"
authors = ["Joan Sebastian Tamayo Rivera", "Angie Marilyn Largo Guzman"]
license = "Proprietary"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.35.11"
pydantic = "^2.8.2"
flatten-json = "^0.1.14"
xmltodict = "^0.13.0"
python-dotenv = "^1.0.1"
pymemcache = "^4.0.0"
psycopg2-binary = "^2.9.9"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"

[tool.poetry.dev-dependencies]
[tool.semantic_release]
version_toml = [
    "pyproject.toml:tool.poetry.version",
]                                                    # version location
branch = "main"                                      # branch to make releases of
changelog_file = "CHANGELOG.md"                      # changelog file
build_command = "pip install poetry && poetry build" # build dists 

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
